# services/payments/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:3004
info:
  contact:
    email: support@theraclosure.com
    name: API Support
    url: http://www.theraclosure.com/support
  description: A comprehensive payments service built with Go, providing customer
    management, subscription handling, and Stripe integration.
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Payments Service API
  version: "1.0"
paths:
...
  /api/v1/webhooks/stripe:
    post:
      consumes:
      - application/json
      description: Processes incoming Stripe webhook events for payment updates
      parameters:
      - description: Stripe webhook signature for verification
        in: header
        name: Stripe-Signature
        required: true
        type: string
      - description: Stripe webhook payload
        in: body
        name: webhook
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Webhook processed successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid webhook signature or payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Handle Stripe webhook
      tags:
      - webhooks
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"